More jacrev bugs

+++ find_plot_eig_all.py
@@ -316,15 +316,15 @@ num_workers = 0  #What does this do?
 
 
 # FNO models and predictions
-mynet_directstep_FNO = FNO1d(modes, width, time_future, time_history).double()
+mynet_directstep_FNO = FNO1d(modes, width, time_future, time_history).float()
 mynet_directstep_FNO.load_state_dict(torch.load('NN_FNO_Directstep_lead1.pt'))
 mynet_directstep_FNO.cuda()
 
-mynet_Eulerstep_FNO = FNO1d(modes, width, time_future, time_history).double()
+mynet_Eulerstep_FNO = FNO1d(modes, width, time_future, time_history).float()
 mynet_Eulerstep_FNO.load_state_dict(torch.load('NN_FNO_Eulerstep_lead1.pt'))
 mynet_Eulerstep_FNO.cuda()
 
-mynet_PECstep_FNO = FNO1d(modes, width, time_future, time_history).double()
+mynet_PECstep_FNO = FNO1d(modes, width, time_future, time_history).float()
 mynet_PECstep_FNO.load_state_dict(torch.load('NN_FNO_PECstep_lead1.pt'))
 mynet_PECstep_FNO.cuda()
 
@@ -332,9 +332,9 @@ val_dict_direct_FNO = scipy.io.loadmat('/media/volume/sdb/conrad_stability/model
 val_dict_Euler_FNO = scipy.io.loadmat('/media/volume/sdb/conrad_stability/model_eval_FNO/predicted_Eulerstep_1024_FNO_lead1.mat')
 val_dict_PEC_FNO = scipy.io.loadmat('/media/volume/sdb/conrad_stability/model_eval_FNO/predicted_PECstep_1024_FNO_lead1.mat')
 
-pred_direct_FNO = torch.tensor(val_dict_direct_FNO[u'prediction'], dtype=torch.double)
-pred_Euler_FNO = torch.tensor(val_dict_Euler_FNO[u'prediction'], dtype=torch.double)
-pred_PEC_FNO = torch.tensor(val_dict_PEC_FNO[u'prediction'], dtype=torch.double)
+pred_direct_FNO = torch.tensor(val_dict_direct_FNO[u'prediction'], dtype=torch.float)
+pred_Euler_FNO = torch.tensor(val_dict_Euler_FNO[u'prediction'], dtype=torch.float)
+pred_PEC_FNO = torch.tensor(val_dict_PEC_FNO[u'prediction'], dtype=torch.float)
 
 ygrad_direct_FNO = torch.zeros([int(4),input_size,input_size])
 ygrad_Euler_FNO = torch.zeros([int(4),input_size,input_size])
@@ -342,15 +342,15 @@ ygrad_PEC_FNO = torch.zeros([int(4),input_size,input_size])
 
 
 # FNO + tendency models and predictions
-mynet_directstep_FNO_tendency = FNO1d(modes, width, time_future, time_history).double()
+mynet_directstep_FNO_tendency = FNO1d(modes, width, time_future, time_history).float()
 mynet_directstep_FNO_tendency.load_state_dict(torch.load('NN_Spectral_Loss_FNO_Directstep_tendency_lambda_reg5lead1.pt'))
 mynet_directstep_FNO_tendency.cuda() 
 
-mynet_Eulerstep_FNO_tendency = FNO1d(modes, width, time_future, time_history).double()
+mynet_Eulerstep_FNO_tendency = FNO1d(modes, width, time_future, time_history).float()
 mynet_Eulerstep_FNO_tendency.load_state_dict(torch.load('NN_Spectral_Loss_FNO_Eulerstep_tendency_lambda_reg5lead1.pt'))
 mynet_Eulerstep_FNO_tendency.cuda()
 
-mynet_PECstep_FNO_tendency = FNO1d(modes, width, time_future, time_history).double()
+mynet_PECstep_FNO_tendency = FNO1d(modes, width, time_future, time_history).float()
 mynet_PECstep_FNO_tendency.load_state_dict(torch.load('NN_Spectral_Loss_FNO_PECstep_tendency_lambda_reg5lead1.pt'))
 mynet_PECstep_FNO_tendency.cuda()
 
@@ -358,9 +358,9 @@ val_dict_direct_FNO_tendency = scipy.io.loadmat('/media/volume/sdb/conrad_stabil
 val_dict_Euler_FNO_tendency = scipy.io.loadmat('/media/volume/sdb/conrad_stability/model_eval_FNO_tendency/predicted_Eulerstep_1024_FNO_tendency_lead1.mat')
 val_dict_PEC_FNO_tendency = scipy.io.loadmat('/media/volume/sdb/conrad_stability/model_eval_FNO_tendency/predicted_PECstep_1024_FNO_tendency_lead1.mat')
 
-pred_direct_FNO_tendency = torch.tensor(val_dict_direct_FNO_tendency[u'prediction'], dtype=torch.double)
-pred_Euler_FNO_tendency = torch.tensor(val_dict_Euler_FNO_tendency[u'prediction'], dtype=torch.double)
-pred_PEC_FNO_tendency = torch.tensor(val_dict_PEC_FNO_tendency[u'prediction'], dtype=torch.double)
+pred_direct_FNO_tendency = torch.tensor(val_dict_direct_FNO_tendency[u'prediction'], dtype=torch.float)
+pred_Euler_FNO_tendency = torch.tensor(val_dict_Euler_FNO_tendency[u'prediction'], dtype=torch.float)
+pred_PEC_FNO_tendency = torch.tensor(val_dict_PEC_FNO_tendency[u'prediction'], dtype=torch.float)
 
 ygrad_direct_FNO_tendency = torch.zeros([int(4),input_size,input_size])
 ygrad_Euler_FNO_tendency = torch.zeros([int(4),input_size,input_size])
@@ -394,8 +394,9 @@ for j in np.array([0, 10000, 50000, 100000]):
     ygrad_direct_FNO_tendency[i,:,:] = torch.func.jacrev(directstep, argnums=1)(mynet_directstep_FNO_tendency, torch.reshape(pred_direct_FNO_tendency[0,:],(1,input_size,1)))
     ygrad_Euler_FNO_tendency[i,:,:] = torch.func.jacrev(Eulerstep, argnums=1)(mynet_Eulerstep_FNO_tendency, torch.reshape(pred_Euler_FNO_tendency[0,:],(1,input_size,1)))
     ygrad_PEC_FNO_tendency[i,:,:] = torch.func.jacrev(PECstep, argnums=1)(mynet_PECstep_FNO_tendency, torch.reshape(pred_PEC_FNO_tendency[0,:],(1,input_size,1)))
-
     i += 1
+
+
 print('All jacobians calculated')
 fig1, ax1 = plt.subplots(figsize=(10,8))
 fig2, ax2 = plt.subplots(figsize=(10,8))
